PROJ [MAINT]: Overhaul the project for better long-term maintainability.
I do realize that individual commits are supposed to be small changes,
but with Rust, if you want to do an overhaul, you gotta do the entire
thing at once before it will allow you to compile the project.

Here is a list of the most significant changes:

    - Removed the concept of "cycle timestamps" in memory accesses.
    - Devices now get queried after each CPU instruction is processed.
    - The curses-based TUI and SDL interface now live in separate threads.
    - Cleaned up internal cross-module function accesses.
    - Separated the machine emulation code from the front-end code.
    - The log crate is used for logging, instead of a custom trait.

None of these changes should be visible to the user.